linter and test issues
